flat build files
