+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "ngx-trend",
-  "version": "2.0.0-beta",
+  "version": "2.0.0-beta.3",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "rimraf dist && node build.js",
+    "build": "rimraf dist && ts-node build.ts",
     "publish": "cd dist/packages-dist && npm publish",
     "bundlesize": "bundlesize",
     "test": "ng test",
@@ -35,14 +35,16 @@
     "@angular/compiler-cli": "^4.3.6",
     "@angular/language-service": "^4.3.6",
     "@types/jasmine": "^2.5.54",
-    "@types/jasminewd2": "^2.0.2",
-    "@types/node": "^8.0.25",
-    "bundlesize": "^0.14.1",
+    "@types/jasminewd2": "^2.0.3",
+    "@types/node": "^8.0.26",
+    "bundlesize": "0.14.2",
     "codelyzer": "^3.1.2",
+    "copy": "^0.3.0",
+    "copyfiles": "^1.2.0",
     "fs-extra": "^4.0.1",
     "jasmine-core": "^2.8.0",
     "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^1.7.0",
+    "karma": "^1.7.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^1.0.1",
     "karma-coverage-istanbul-reporter": "^1.3.0",
@@ -50,14 +52,14 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "^5.1.2",
     "rimraf": "^2.6.1",
-    "rollup": "^0.49.1",
+    "rollup": "^0.49.2",
     "rollup-plugin-cleanup": "^1.0.1",
     "rollup-plugin-filesize": "^1.4.2",
     "rollup-plugin-node-resolve": "^3.0.0",
     "rollup-plugin-sourcemaps": "^0.4.2",
     "ts-node": "^3.3.0",
     "tslint": "^5.7.0",
-    "typescript": "^2.4.2"
+    "typescript": "^2.5.2"
   },
   "bundlesize": [
     {
