+++ src/app/app.component.spec.ts
@@ -5,7 +5,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NouisliderModule } from 'ng2-nouislider';
 
 import { AppComponent } from './app.component';
-import { TrendModule } from '../lib';
+import { TrendModule } from '../lib/trend';
 import { ConfigComponent } from './config/config.component';
 import { FooterComponent } from './footer/footer.component';
 import { GithubLinkComponent } from './github-link/github-link.component';
