+++ src/app/app.module.ts
@@ -4,8 +4,9 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { NouisliderModule } from 'ng2-nouislider';
 
+import { TrendModule } from '../lib/trend';
+
 import { AppComponent } from './app.component';
-import { TrendModule } from '../lib';
 import { ConfigComponent } from './config/config.component';
 import { FooterComponent } from './footer/footer.component';
 import { GithubLinkComponent } from './github-link/github-link.component';
